Move Calculator classes into their own directory
